+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/PairedResultGraph/BasicPairedResultGraphs/SimplePairedResult.jsx
@@ -4,14 +4,14 @@ import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import '@cerner/terra-graphs/lib/components/Graph.module.scss';
 import '@cerner/terra-graphs/lib/components/PairedResult/PairedResultGraph.module.scss';
-import pairedResultDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultDefault';
+import getPairedResultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultDefault';
 import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/PairedResult/pairedResultSimpleData';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(pairedResultDefault('#simplePairedResultGraph'));
-const dataset = utils.deepClone(exampleData);
+const graphConfig = utils.deepClone(getPairedResultConfig('#simplePairedResultGraph'));
+const dataset = [utils.deepClone(exampleData)];
 
 const SimplePiaredResultExample = () => {
   React.useEffect(() => {