+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/PairedResultGraph/BasicPairedResultGraphs/TimeseriesWithEventline.jsx
@@ -4,33 +4,14 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import '@cerner/terra-graphs/lib/components/Graph.module.scss';
 import '@cerner/terra-graphs/lib/components/PairedResult/PairedResultGraph.module.scss';
-import pairedResultTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultTimeseries';
+import getPairedResultEventlineTimeseriesConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultTimeseriesWithEventline';
 import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/PairedResult/pairedResultTimeseriesData';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(pairedResultTimeseries('#pairedResultWithEventline'));
-graphConfig.eventline = [
-  {
-    color: Carbon.helpers.COLORS.GREY,
-    style: {
-      strokeDashArray: '4,4',
-    },
-    value: new Date(2016, 8, 1, 8).toISOString(),
-  },
-  {
-    color: Carbon.helpers.COLORS.BLACK,
-    style: {
-      strokeDashArray: '2,2',
-    },
-    value: new Date(2017, 5, 1, 12).toISOString(),
-  },
-];
-graphConfig.clickPassThrough = {
-  dateline: false,
-};
-const dataset = utils.deepClone(exampleData);
+const graphConfig = utils.deepClone(getPairedResultEventlineTimeseriesConfig('#pairedResultWithEventline'));
+const dataset = [utils.deepClone(exampleData)];
 
 const PiaredResultWithEventlineExample = () => {
   React.useEffect(() => {