+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/PairedResultGraph/Reflow/DynamicEventline.jsx
@@ -4,26 +4,17 @@ import PairedResultGraph from '@cerner/terra-graphs/lib/components/PairedResult/
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import getPairedResultTimeseriesPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultPanning';
+import getPairedResultTimeseriesPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/panningWithEventline';
 import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/PairedResult/simplePanningData';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(getPairedResultTimeseriesPanningConfig('#PairedResultDynamicEventline'));
-graphConfig.eventline = [
-  {
-    color: Carbon.helpers.COLORS.GREY,
-    style: {
-      strokeDashArray: '4,4',
-    },
-    value: new Date(2016, 0, 1, 8).toISOString(),
-  },
-];
 const initialState = {
   initial: 0,
   factor: 3,
-  dataset: [utils.deepClone(data[0])],
+  dataset: [utils.deepClone(data)],
   graphConfig,
 };
 
@@ -31,7 +22,7 @@ const PairedResultPanningExample = () => {
   const reducer = (panState, action) => {
     const newGraphState = utils.deepClone(panState.graphConfig);
     let hour;
-    const newDataset = [utils.deepClone(data[0])];
+    const newDataset = [utils.deepClone(data)];
 
     switch (action.type) {
       case 'panLeft':